refactor: ♻️ packages to libs
