+++ libs/graphql-code-generator/package.json
