+++ libs/graphql-code-generator/project.json
@@ -1,7 +1,7 @@
 {
   "name": "graphql-code-generator",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/graphql-code-generator/src",
+  "sourceRoot": "libs/graphql-code-generator/src",
   "projectType": "library",
   "tags": [],
   "targets": {
