+++ libs/graphql-code-generator/tsconfig.lib.json
