+++ libs/graphql-code-generator/tsconfig.spec.json
