+++ libs/react-markdown-mermaid/package.json
