+++ libs/react-markdown-mermaid/project.json
@@ -1,7 +1,7 @@
 {
   "name": "react-markdown-mermaid",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/react-markdown-mermaid/src",
+  "sourceRoot": "li b s/react-markdown-mermaid/src",
   "projectType": "library",
   "tags": [],
   "targets": {
@@ -9,17 +9,17 @@
       "executor": "@nx/vite:build",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/packages/react-markdown-mermaid",
-        "main": "packages/react-markdown-mermaid/src/index.ts",
-        "tsConfig": "packages/react-markdown-mermaid/tsconfig.lib.json",
-        "assets": ["packages/react-markdown-mermaid/*.md"]
+        "outputPath": "dist/libs/react-markdown-mermaid",
+        "main": "libs/react-markdown-mermaid/src/index.ts",
+        "tsConfig": "libs/react-markdown-mermaid/tsconfig.lib.json",
+        "assets": ["libs/react-markdown-mermaid/*.md"]
       }
     },
     "release": {
       "executor": "nx:run-commands",
       "outputs": [],
       "options": {
-        "command": "npx semantic-release-plus --extends ./packages/react-markdown-mermaid/release.config.cjs",
+        "command": "npx semantic-release-plus --extends ./libs/react-markdown-mermaid/release.config.cjs",
         "parallel": false
       }
     }
