+++ libs/react-markdown-mermaid/tsconfig.lib.json
